move curryToArity function to separate module, add some tests, fix linter
