+++ .eslintrc.json
@@ -1,11 +1,17 @@
 {
-    "extends": ["airbnb-base", "prettier"],
-    "plugins": ["prettier"],
-    "rules": {
-        "prettier/prettier": "error"
-    },
-    "env": {
-        "node": true,
-        "jest": true
-    }
+  "extends": [
+    "airbnb-base",
+    "prettier"
+  ],
+  "plugins": [
+    "prettier"
+  ],
+  "rules": {
+    "prettier/prettier": "error",
+    "import/extensions": "off"
+  },
+  "env": {
+    "node": true,
+    "jest": true
+  }
 }
