+++ package.json
@@ -11,6 +11,7 @@
   },
   "scripts": {
     "build": "microbundle src/index.js",
+    "lint": "eslint src",
     "test": "NODE_OPTIONS=--experimental-vm-modules jest src",
     "prepublishOnly": "npm run build"
   },
