+++ src/index.js
@@ -1,16 +1,4 @@
-// curry a function to a specified arity
-const curryToArity = (fn, arity) => {
-  const resolver =
-    (...args) =>
-    (...innerArgs) => {
-      // make sure the function won't return functions endlessly when called with no arguments
-      const newArgs =
-        arity !== 0 && innerArgs.length === 0 ? [undefined] : innerArgs;
-      const allArgs = [...args, ...newArgs];
-      return allArgs.length >= arity ? fn(...allArgs) : resolver(...allArgs);
-    };
-  return resolver();
-};
+import curryToArity from "./util.js";
 
 /* Small library for combining matching functions in a reusable and
  * readable way
